Computing grad for sum and mul operations.
A complete definition of the `Value` class that wraps
a `FloatingPoint` literal. As well, we introduce operations
for sum and mul, including their forward and backward passes.
Those binary operations support as well one of either arguments
to be a `FloatingPoint` literail, for instane, sum:

```swift
let l = 3.0 + Value(2.0)
```

We introduce as well a basic suite but complete suit of tests.