Fixing order in gradient propagation.
As topological sort will compute the inversed order on how
to traverse the nodes of the computation graph to propagate
the gradients from the loss function back to the inputs, we
need to reverse the result of the sort.

As well, we extend the suite of tests with a more complicated
expression that involves mult and sum operations.